Extract cmdInit.ApplyConfig into a separete initApplyConfig function
No logic change, just moving around code.
